Fix tests etc. for explicit numba signatures

+++ pax/plugins/ZLE.py
@@ -43,7 +43,8 @@ class SoftwareZLE(plugin.TransformPlugin):
 
             # Find intervals above ZLE threshold
             high_threshold = low_threshold = self.config['zle_threshold']
-            n_itvs_found = find_intervals_above_threshold(w, high_threshold, low_threshold, zle_intervals_buffer)
+            n_itvs_found = find_intervals_above_threshold(w.astype(np.float64),
+                                                          high_threshold, low_threshold, zle_intervals_buffer)
 
             if n_itvs_found == self.config['max_intervals']:
                 # more than 5000 intervals - insane!!!
